feat: refactor folder name
